Added code for stretch project
The stretch project is a similar templating strategy as that proposed
in the course to fine model but applied to the data inputs. This templating
strategy replaces all data tokens in the input with template variables in a
deterministic manner. This is done to help reduce hallucinations in the generated
summaries and improve performance.